Refactor frontend components
Refactor Hero, AssetManagementSection, WaitlistModal, and carousel components into smaller, more focused components to improve readability, reusability, and maintainability.
